refactoring: switch to throw, move arma bits to separate file
