+++ src/partitions.h
@@ -66,7 +66,7 @@ inline std::vector<size_t> in_terms_positions(const TermMap& labels_map,
   for (const auto& term : term_summary.unique_terms) {
     auto it = labels_map.find(term);
     if (it == labels_map.end())
-      stop("Internal error: term not found in labels_input");
+      throw std::runtime_error("term not found in labels_input");
     needed.push_back(it->second);
   }
   return needed;
